+++ CHANGELOG.md
@@ -16,6 +16,7 @@ For a full diff see [`fa9c564...main`][fa9c564...main].
 * Added `FieldDefinition::closure()` which allows resolving a field definition to the return value of a closure that is invoked with the `FixtureFactory` ([#155]), by [@localheinz]
 * Allowed creation of optional field definitions ([#167]) and ([#196]), by [@localheinz]
 * Allowed using field definitions as field overrides ([#270]), by [@localheinz]
+* Allowed use of `Faker\Generator` in field definitions ([#144]), by [@localheinz]
 
 ### Changed